fix some bug BE

+++ src/main/java/com/iuh/IUHStudent/response/diem/DiemMonHoc.java
@@ -9,8 +9,8 @@ public class DiemMonHoc {
     private String tenMonHoc;
     private String ghiChu;
     private List<Double> diemThuongKy;
-    private Float diemGiuaKy;
+    private Double diemGiuaKy;
     private List<Double> diemThucHanh;
-    private Float diemCuoiKy;
+    private Double diemCuoiKy;
     private Double diemTrungBinh;
 }
